Microservices_video_detection-main/outbox/requirements.txt
fastapi
uvicorn
httpx
pymongo
kafka-python
